use crackle snapshot + start making changes
