+++ .eslintrc.js
@@ -43,7 +43,7 @@ module.exports = {
           {
             patterns: [
               'braid-design-system**',
-              '!braid-design-system/lib/utils/source.macro',
+              '!braid-design-system/src/lib/utils/source.macro',
               'braid-src/**',
               'site/**',
               '**/site/**',
