+++ .prettierrc
@@ -1,5 +1,13 @@
 {
   "singleQuote": true,
   "tabWidth": 2,
-  "trailingComma": "all"
+  "trailingComma": "all",
+  "overrides": [
+    {
+      "files": "**/scripts/*",
+      "options": {
+        "printWidth": 120
+      }
+    }
+  ]
 }
