+++ packages/braid-design-system/crackle.config.ts
@@ -1,0 +1,9 @@
+import type { CrackleConfig } from '@crackle/core';
+
+const config: CrackleConfig = {
+  // dtsOptions: {
+  //   noEmitOnError: true,
+  // }
+};
+
+export default config;
