+++ packages/braid-design-system/playroom/components.ts
@@ -14,4 +14,4 @@ if (global?.__IS_PLAYROOM_ENVIRONMENT__ !== 'clearly') {
   `);
 }
 
-export * from '../lib/playroom/components';
+export * from '../src/lib/playroom/components';
