+++ packages/braid-design-system/sku.config.js
@@ -1,16 +1,7 @@
 const browserslist = require('../../browserslist');
 
 module.exports = {
-  srcPaths: [
-    './color-mode',
-    './css',
-    './lib',
-    './reset',
-    './scripts',
-    './test',
-    './themes',
-  ],
-  clientEntry: './lib/components/index.ts',
+  clientEntry: './src/lib/components/index.ts',
   rootResolution: false,
   target: './dist',
   supportedBrowsers: browserslist,
