+++ packages/codemod/src/fixtures/testFixture.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-restricted-imports
-import { atoms } from 'braid-src/css';
+import { atoms } from 'braid-src/entries/css';
 // @ts-expect-error
 import { myShadow } from './shadows';
