+++ site/playroom.config.js
@@ -3,17 +3,17 @@ const SkuWebpackPlugin = require('sku/webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const browserslist = require('../browserslist');
 
-const braidSrc = '../packages/braid-design-system';
+const braidRoot = path.join(__dirname, '../packages/braid-design-system');
+const resolveFromBraid = (p) => require.resolve(path.join(braidRoot, p));
 
 module.exports = {
-  title: 'BRAID',
   outputPath: './dist/playroom',
-  components: `${braidSrc}/lib/playroom/components.ts`,
-  snippets: `${braidSrc}/lib/playroom/snippets.ts`,
-  themes: `${braidSrc}/lib/themes/index.ts`,
-  frameComponent: `${braidSrc}/lib/playroom/FrameComponent.tsx`,
-  scope: `${braidSrc}/lib/playroom/useScope.ts`,
-  typeScriptFiles: [`${braidSrc}/lib/**/*.{ts,tsx}`, '!**/node_modules'],
+  components: resolveFromBraid('src/lib/playroom/components.ts'),
+  snippets: resolveFromBraid('src/lib/playroom/snippets.ts'),
+  themes: resolveFromBraid('src/lib/themes/index.ts'),
+  frameComponent: resolveFromBraid('src/lib/playroom/FrameComponent.tsx'),
+  scope: resolveFromBraid('src/lib/playroom/useScope.ts'),
+  typeScriptFiles: [`${braidRoot}/src/**/*.{ts,tsx}`],
   widths: [320, 768, 1024, 1400],
   openBrowser: false,
   port: 8082,
@@ -23,9 +23,7 @@ module.exports = {
         ignoreOrder: true,
       }),
       new SkuWebpackPlugin({
-        include: ['lib', 'css', 'reset', 'themes'].map((dir) =>
-          path.join(__dirname, braidSrc, dir),
-        ),
+        include: [path.join(braidRoot, 'src')],
         target: 'browser',
         browserslist,
         mode: process.env.NODE_ENV ? process.env.NODE_ENV : 'development',
