+++ site/src/App/routes/examples/marketing-banner/marketing-banner.tsx
@@ -20,7 +20,7 @@ import { Placeholder } from 'braid-src/lib/playroom/components';
 import Code from '../../../Code/Code';
 import type { ReactNodeNoStrings } from 'braid-src/lib/components/private/ReactNodeNoStrings';
 import { PageTitle } from '../../../Seo/PageTitle';
-import source from 'braid-design-system/lib/utils/source.macro';
+import source from 'braid-design-system/src/lib/utils/source.macro';
 import { LinkableHeading } from '../../../LinkableHeading/LinkableHeading';
 
 interface StepProps {
