+++ site/src/App/routes/gallery/Gallery.tsx
@@ -44,9 +44,9 @@ import {
 // TODO: COLORMODE RELEASE
 // Use public import
 import { Box } from 'braid-src/lib/components/Box/Box';
-import docsTheme from 'braid-src/themes/docs';
+import docsTheme from 'braid-src/entries/themes/docs';
 import { getHistory, isNew } from '../../Updates';
-import source from 'braid-design-system/lib/utils/source.macro';
+import source from 'braid-design-system/src/lib/utils/source.macro';
 import { CodeButton, formatSnippet } from '../../Code/Code';
 import type { ComponentExample } from '../../../types';
 import {
