+++ site/src/App/routes/guides/development-workflow.tsx
@@ -17,7 +17,7 @@ import { TextStack } from '../../TextStack/TextStack';
 import Code from '../../Code/Code';
 import type { Page } from '../../../types';
 import { PageTitle } from '../../Seo/PageTitle';
-import source from 'braid-design-system/lib/utils/source.macro';
+import source from 'braid-design-system/src/lib/utils/source.macro';
 import { InlineCode } from '../../InlineCode/InlineCode';
 import { LinkableHeading } from '../../LinkableHeading/LinkableHeading';
