+++ site/src/render.tsx
@@ -12,7 +12,7 @@ import * as themes from 'braid-src/lib/themes';
 import { braidVersionToDate } from './getVersionDetails';
 import { initUpdates } from './App/Updates';
 import packageJson from 'braid-design-system/package.json';
-import { __experimentalDarkMode__ } from 'braid-src/color-mode';
+import { __experimentalDarkMode__ } from 'braid-src/entries/color-mode';
 
 const { version } = packageJson;
